Added output information API.
Added a detailed output info API. This includes the current color of the output (as RGB array and hex string), the output number and number of layers, as well as a detailed specification of each layer. This resolves #4.

For each layer, the report includes the color (as RGB array and hex string), opacity, transition envelope and blend function.

An example of the output (with only the bottom layer set):

```json
{
  "mixedColorHex": "#3eb4ce",
  "layers": [
    {
      "colorHex": "#6acfe3",
      "opacity": 1,
      "colorRgb": [106, 207, 227],
      "envelope": "CosineEnvelope",
      "blender": "LabAverage"
    },
    {
      "colorHex": "#000000",
      "opacity": 0,
      "colorRgb": [0, 0, 0],
      "envelope": "CosineEnvelope",
      "blender": "LabAverage"
    },
    {
      "colorHex": "#000000",
      "opacity": 0,
      "colorRgb": [0, 0, 0],
      "envelope": "CosineEnvelope",
      "blender": "LabAverage"
    }
  ],
  "outputNumber": 0,
  "layerCount": 3,
  "mixedColorRgb": [62, 180, 206]
}
```