Moved API landing page to separate file.
The file is currently read on each API request, which is not ideal but the number of hits on this particular page is expected to be very limited. The page now contains a little information and links to the GitHub page, Frack wiki and other relevant things. This is a first step to satisfying issue #5.