Tripled API live view performance
Changed the API call to retrieve output information to a single request to get the state for all outputs. This significantly reduces the number of JSON requests that have to be performed to generate a full live view of the Lightbox system. In the tested browser, this leads to a threefold reduction in CPU load when the API view is active.