Added Arduino code for Lightbox.
Added the Arduino equivalent of the JTagController as it was originally created for the Twitterboom project. The original AVR code has been rewritten in a minimal Arduino sketch which has the exact same behavior.

There is one small catch, to get the best performance, the Wire library must be changed, specifically the I2C frequency should be set to 400kHz from the default 100kHz. For this, copy the Wire library to the 'sketchbook/libraries' directory and open the 'Wire/utility/twi.h' file. The TWI_FREQ needs to be adjusted from 100000L to 400000L. This improves the performance of the controller from ~112Hz to ~200Hz.