Added new Arduino controller sketch.
The NewController is a redesign of the JTagController which has fixed-length binary commands. This reduces the command length for a "set channel to white" from 16 to 6 characters. Beyond the length reduction, the new controller does not send confirmations when a command has been received, greatly improving performance and reducing complexity. Heartbeats are also removed, as long as commands are received at least once per ten seconds, the controller will keep the outputs engaged.