Simplify Command method by always executing _Verify
Controllers that implement verification can implement the _Verify method, those that don't shouldn't.

This removes an if statement, simplifying logic and removing necessary class variables. Commands that need to be sent without verification (such as the heartbeat on the JTagController) can send commands directly to _Command. The assumption is made that only class internals will need to send unverified commands in an otherwise verifying controller.