Create PCA9685 RGB LED class.
Removed all the direct Wire line twiddling from the main controller sketch so that the commands to drive the PCA9685 controller can be re-used for other hardware controllers as well. Moved gamma correction and the like into that library as well.

There are a number of bugs still present in the library. For one, setAll doesn't work, likely because either the Wire transmit buffer is not large enough to accommodate the data, or the PCA9685 doesn't like it when many bytes are sent. Currently, using the setAll method results in 2 LEDs responding, and the red of the third LED _might_ respond.