Updated PCA9685 controller class
Public functions that set brightness of LEDs now do so three pins at a time, to avoid transmit buffer overflow when writing.

Added PWM start offsets. The fifth LED will start its PWM cycle nearer the end of the 4096 step cycle. Ideally, this evens the load on the connected power supply. There appear to be a few bugs related to it, where under some high brightness levels, where the PWM level is significantly lower than expected.

Added function declarations to the top of the sketch, this helps with commandline compiling the sketch without re-ordering the functions.