Introduced transparency checkerboards for colors and Added jQuery UI.
For layers that are set to be partially transparent, a simple checkerboard has been introduced so that the transparency of the color can be distinguished from a lighter tint of the same color. jQuery UI has been added to provide sliders for the opacity and steps selection.

The current theme selection is 'smoothness' which provides a plain but elegant sliders.