Made sliders functional and improved popup window.
The colorpicker in the popup window is immediately placed, before the lightbox popup is triggered. The current color of the picker is set in the lightbox done callback, as before that, the picker doesn't properly update the indicator positions.

Added callbacks for the opacity and steps sliders, creating visual feedback on the current values. Changing the opacity also immediately send out commands to the server (similar to the color picker updates).